generter修改升级

+++ springbootstudydomain/src/main/java/com/baili/dao/MemberMapper.java
@@ -3,12 +3,11 @@ package com.baili.dao;
 import com.baili.entity.Member;
 import com.baili.entity.MemberCriteria;
 import java.util.List;
-
 import org.apache.ibatis.annotations.Param;
 import org.apache.ibatis.session.RowBounds;
 
 public interface MemberMapper {
-    long countByExample(MemberCriteria example);
+    int countByExample(MemberCriteria example);
 
     int deleteByExample(MemberCriteria example);
